requirements.txt
discord.py
pytz
asyncio